chore: bump axios to 0.21.0 where we can
botkit-mock pins the version of botbuilder it depends on, so we'll ahve
to wait for a new version of botkit-mock before we get a vulnerable
version of axios entirely out of the codebase. But this is still worth
landing, I think, because it gets it out of the actual code for the app.
The vulnerable version is used only in the test mock with this change.